+++ backend/src/main/java/org/example/backend/jwt/JWTFilter.java
@@ -11,8 +11,8 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.List;
 import lombok.extern.slf4j.Slf4j;
-import org.example.backend.admin.domain.entity.Admin;
-import org.example.backend.admin.domain.entity.CustomUserDetails;
+import org.example.backend.users.domain.entity.Admin;
+import org.example.backend.users.domain.entity.CustomUserDetails;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
