+++ backend/src/main/java/org/example/backend/users/domain/dto/AdminReqDto.java
@@ -1,12 +1,15 @@
-package org.example.backend.admin.domain.dto;
+package org.example.backend.users.domain.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-public class AdminResDto {
+public class AdminReqDto {
 
     private String loginId;
+
+    private String password;
+
     private String email;
 }
