+++ backend/src/main/java/org/example/backend/users/domain/dto/SignInResDto.java
@@ -1,7 +1,6 @@
-package org.example.backend.admin.domain.dto;
+package org.example.backend.users.domain.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
