+++ backend/src/main/java/org/example/backend/users/domain/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package org.example.backend.admin.domain.dto;
+package org.example.backend.users.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
