--- backend/src/main/java/org/example/backend/users/dto/SjUserProfile.java
@@ -1,19 +1,0 @@
-package org.example.backend.users.dto;
-
-public class SjUserProfile {
-    private final String name;
-    private final String major;
-
-    public SjUserProfile(String name, String major) {
-        this.name = name;
-        this.major = major;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getMajor() {
-        return major;
-    }
-}
