+++ backend/src/main/java/org/example/backend/users/service/AdminService.java
@@ -1,15 +1,15 @@
-package org.example.backend.admin.service;
+package org.example.backend.users.service;
 
-import static org.example.backend.admin.exception.AdminExceptionType.ALREADY_EXIST_LOGIN_ID;
-import static org.example.backend.admin.exception.AdminExceptionType.INVALID_ACCESS_TOKEN;
+import static org.example.backend.users.exception.AdminExceptionType.ALREADY_EXIST_LOGIN_ID;
+import static org.example.backend.users.exception.AdminExceptionType.INVALID_ACCESS_TOKEN;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.backend.admin.domain.dto.AccessTokenReq;
-import org.example.backend.admin.domain.dto.SignInReqDto;
-import org.example.backend.admin.domain.entity.Admin;
-import org.example.backend.admin.exception.AdminException;
-import org.example.backend.admin.repository.AdminRepository;
+import org.example.backend.users.domain.dto.AccessTokenReq;
+import org.example.backend.users.domain.dto.SignInReqDto;
+import org.example.backend.users.domain.entity.Admin;
+import org.example.backend.users.exception.AdminException;
+import org.example.backend.users.repository.AdminRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
